fix: Fix hot-reloading
Previously, the export default of Root was causing issues with
hot-loader.  Since the file was being imported with a default export it
couldn't replace the updated Root function, it would have to re-import
the entire file which caused a full reload each time.

Additionally, the ThemeProvider was moved to Root since it was
unnecessarily being re-rendered each time app updated, when it doesn't
need to change once initializied.